server/requirements.txt
# Install with: `pip install -r requirements.txt`
 
Django==1.11
django-admin-honeypot==1.0.0
django-cors-headers==2.0.2
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
gunicorn==19.7.1
markdown2==2.3.1
numpy==1.13.1
olefile==0.44
pandas==0.20.3
Pillow==4.2.1
PyJWT==1.4.0
PyMySQL==0.7.11
python-dateutil==2.6.0
pytz==2017.2
scikit-learn==0.18.2
scipy==0.19.1
six==1.10.0
SQLAlchemy==1.1.11
bokeh==0.12.6
xlrd==1.0.0
bkcharts==0.2
selenium

# Install this thing
# sudo apt-get install phantomjs


#pandas
sklearn
#json
#scipy
#xlrd
#bokeh
#bkcharts
# >>>>>>> Stashed changes
pywebhdfs
